+++ src/Symfony/Bridge/Doctrine/Form/Type/DoctrineType.php
@@ -150,7 +150,8 @@ abstract class DoctrineType extends AbstractType implements ResetInterface
                     $options['em'],
                     $options['class'],
                     $options['id_reader'],
-                    $entityLoader
+                    $entityLoader,
+                    false
                 );
 
                 if (null !== $hash) {